Remove resource module, fixes for client code, support destroy action
