+++ gcp/modules/gcp-stackdriver-alerting/resources/notification_channels/sergey_email.json
